+++ etils/ecolab/__init__.py
@@ -17,6 +17,7 @@
 from etils.ecolab.array_as_img import auto_plot_array
 from etils.ecolab.colab_utils import collapse
 from etils.ecolab.colab_utils import json
+from etils.ecolab.inspects.core import inspect
 from etils.ecolab.module_utils import clear_cached_modules
 from etils.ecolab.patch_utils import patch_graphviz
 from etils.ecolab.patch_utils import set_verbose
@@ -25,6 +26,7 @@ __all__ = [
     'auto_plot_array',
     'collapse',
     'clear_cached_modules',
+    'inspect',
     'patch_graphviz',
     'set_verbose',
 ]