+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changelog follow https://keepachangelog.com/ format.
     *   Add `skip=['jnp', ...]` kwarg to `enp.testing.parametrize_xnp` to
         exclude a specific xnp module from tests.
 *   `ecolab`:
+    *   Add `ecolab.inspect` for interactively inspect any Python objects.
     *   Add `ecolab.json` for interactive expandable JSON display.
     *   `ecolab.auto_plot_array`
         *   Add `torch`